requirements.txt
annotated-types @ file:///C:/b/abs_0dmaoyhhj3/croot/annotated-types_1709542968311/work
anyio @ file:///C:/b/abs_847uobe7ea/croot/anyio_1706220224037/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
cachetools==5.3.3
certifi @ file:///C:/b/abs_35d7n66oz9/croot/certifi_1707229248467/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama @ file:///C:/ci_311/colorama_1676422310965/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h11 @ file:///C:/b/abs_1czwoyexjf/croot/h11_1706652332846/work
httpcore @ file:///C:/b/abs_55n7g233bw/croot/httpcore_1706728507241/work
httplib2==0.22.0
httpx @ file:///C:/b/abs_70wr6jzznw/croot/httpx_1706887145357/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
ipython @ file:///C:/b/abs_b6pfgmrqnd/croot/ipython_1704833422163/work
jedi @ file:///C:/ci_311/jedi_1679427407646/work
matplotlib-inline @ file:///C:/ci_311/matplotlib-inline_1676425798036/work
openai==1.30.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pdf2image==1.17.0
pillow==10.3.0
prompt-toolkit @ file:///C:/b/abs_68uwr58ed1/croot/prompt-toolkit_1704404394082/work
proto-plus==1.23.0
protobuf==4.25.3
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic @ file:///C:/b/abs_f6ffn86wjx/croot/pydantic_1709578044014/work
pydantic_core @ file:///C:/b/abs_05smrkbwjm/croot/pydantic-core_1709574000211/work
Pygments @ file:///C:/b/abs_fay9dpq4n_/croot/pygments_1684279990574/work
pyparsing==3.1.2
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
requests @ file:///C:/b/abs_474vaa3x9e/croot/requests_1707355619957/work
rsa==4.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///C:/b/abs_3akdewudo_/croot/sniffio_1705431337396/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tqdm @ file:///C:/b/abs_551q5__bmw/croot/tqdm_1714567744381/work
traitlets @ file:///C:/ci_311/traitlets_1676423290727/work
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
uritemplate==4.1.1
urllib3 @ file:///C:/b/abs_8e4z8_gh1l/croot/urllib3_1715636317140/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
